requirements.txt
openpyxl==3.1.5
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.1
pandas==2.2.3
